FileLoader results reorganization
Changed sendRequest to return the req object when completed instead of
responseText when success or the whole object when error.

Changed load file to return a results object with a files object that
is indexed by file name and contains the data from the file, and if
there were errors, there will also be an errors object indexed by file
name containing the req object from sendReq.